requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
opencv-python
pillow
PyYAML
ensure
python-box
notebook
setuptools
dvc[gdrive]
-e .